minitalk

+++ client.c
@@ -6,15 +6,15 @@
 /*   By: crachid- <crachid-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/13 20:07:51 by crachid-          #+#    #+#             */
-/*   Updated: 2025/03/20 13:21:05 by crachid-         ###   ########.fr       */
+/*   Updated: 2025/03/24 21:38:45 by crachid-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minitalk.h"
 
-int	ft_isdigit(char *str)
+static int	is_number(char *str)
 {
-	int i;
+	int	i;
 
 	i = 0;
 	while (str[i] != '\0')
@@ -24,72 +24,54 @@ int	ft_isdigit(char *str)
 		i++;
 	}
 	return (1);
-} 
+}
 
-void send_signal(int pid, char c)
+static void	send_signal(int pid, char c)
 {
-	int bit;
+	int	bit;
 
-	bit = 7;
-	while (bit >= 0)
+	bit = 8;
+	while (--bit >= 0)
 	{
 		if ((c >> bit) & 1)
 		{
-			kill(pid, SIGUSR1);
+			if (kill(pid, SIGUSR1) == -1)
+			{
+				ft_printf(2, "Error: kill failed\n");
+				exit(1);
+			}
 		}
 		else
 		{
-			kill(pid, SIGUSR2);
+			if (kill(pid, SIGUSR2) == -1)
+			{
+				ft_printf(2, "Error: kill failed\n");
+				exit(1);
+			}
 		}
 		usleep(100);
-		bit--;
 	}
 }
 
-// void handle_ack(int sig) {
-	// static int received = 0;
-
-	// if (sig == SIGUSR1)
-	//     ++received;
-	// else
-	// {
-	//     ft_putnbr(received);
-	//     ft_putchar('\n');
-	//     exit(0);
-	// }
-	// ft_printf("signal recu avec succes %d", sig);
-// 	(void)sig;
-// }
-
-int main(int argc, char **argv)
+int	main(int argc, char **argv)
 {
-	int pid_server;
-	int i;
+	int	pid_server;
+	int	i;
 
 	i = 0;
-	if(argc != 3)
-	{
-		ft_printf("Error \nUsage: ./client <PID> <message>\n");
-		return (1);
-	}
-	if (!ft_isdigit(argv[1]))
-	{
-		ft_printf("Error: invalid PID\n");
-		return (1);
-	}
+	if (argc != 3)
+		return (ft_printf(2, "Error \nUsage: ./client <PID> <message>\n"), 1);
+	if (!is_number(argv[1]))
+		return (ft_printf(2, "Error: invalid PID\n"), 1);
 	pid_server = ft_atoi(argv[1]);
-
-	if (pid_server <= 0)
+	kill(pid_server, 0);
+	if (pid_server <= 0 || errno == ESRCH)
+		return (ft_printf(2, "Error: invalid PID\n"), 1);
+	while (argv[2][i])
 	{
-		ft_printf("Error: invalid PID\n");
-		return (1);
-	}
-	// signal(SIGUSR1, handle_ack);
-	// signal(SIGUSR2, handle_ack);
-	while (argv[2][i]) {
 		send_signal(pid_server, argv[2][i]);
 		i++;
 	}
 	send_signal(pid_server, '\0');
-	return 0;
+	return (0);
 }
